Admin deletion negation (#64)
Fix for all implemented deletion confirmations and added supporting tests

+++ resources/admin_ui/src/app/pages/composer/index/index.component.ts
@@ -18,7 +18,6 @@ export class IndexComponent implements OnInit {
 
   packages$ !: Observable<Paginated<Package> | null>;
 
-
   constructor(
     private packageService: PackageService,
     private router: Router,
@@ -41,11 +40,13 @@ export class IndexComponent implements OnInit {
     let dialogSubscription = this.dialog
       .open(DeleteConfirmationDialogComponent)
       .afterClosed()
-      .subscribe(result => {
-        let deleteSubscription = this.packageService.delete(pkg.id!).subscribe(() => {
-          this.ngOnInit();
-          setTimeout(() => deleteSubscription.unsubscribe());
-        });
+      .subscribe(confirmation => {
+        if(confirmation) {
+          let deleteSubscription = this.packageService.delete(pkg.id!).subscribe(() => {
+            this.ngOnInit();
+            setTimeout(() => deleteSubscription.unsubscribe());
+          });
+        }
         setTimeout(() => dialogSubscription.unsubscribe());
       });
   }
