+++ resources/admin_ui/src/app/pages/drinks/index/index.component.ts
@@ -40,12 +40,14 @@ export class IndexComponent implements OnInit {
     let dialogSubscription = this.dialog
       .open(DeleteConfirmationDialogComponent)
       .afterClosed()
-      .subscribe(result => {
-        let deleteSubscription = this.drinkService.delete(drink.id!).subscribe(() => {
-          this.ngOnInit();
-          setTimeout(() => deleteSubscription.unsubscribe());
-        });
-        setTimeout(() => dialogSubscription.unsubscribe());
+      .subscribe(confirmation => {
+        if(confirmation) {
+          let deleteSubscription = this.drinkService.delete(drink.id!).subscribe(() => {
+            this.ngOnInit();
+            setTimeout(() => deleteSubscription.unsubscribe(), 0);
+          });
+        }
+        setTimeout(() => dialogSubscription.unsubscribe(), 0);
       });
   }
 }
